refactor: move S3 client initialization to library and remove registry-based credential resolution
